create submodule for separate_objects functions and others
